+++ proptool/checks/base/check.py
@@ -8,17 +8,20 @@
 """
 
 from abc import ABC, abstractmethod
-from typing import Union
+from typing import Dict, Union
 
-from proptool.config import Config
+from proptool.config.config import Config
 from proptool.report.report import Report
 
 
 # noinspection PyUnresolvedReferences
 class Check(ABC):
-    def __init__(self, config: Config):
-        if not isinstance(config, Config):
-            raise ValueError('Invalid type of config passed.')
+    DEFAULT_CHECK_CONFIG = {}
+
+    def __init__(self, config: Union[Config, None] = None):
+        if config is not None:
+            if not isinstance(config, Config):
+                raise ValueError(f'Configuration object must be instance of Config ("{type(config)}" given).')
         self.config = config
 
     @abstractmethod
@@ -31,3 +34,14 @@ class Check(ABC):
             raise ValueError(f'Translation must be valid PropFile object ({type(translation)} given).')
         if reference is None:
             raise ValueError(f'Reference must be valid PropFile object ({type(reference)} given).')
+
+    def need_valid_config(self) -> None:
+        if not isinstance(self.config, Config):
+            raise ValueError(f'Configuration object must be instance of Config ("{type(self.config)}" given).')
+
+    def get_default_config(self) -> Dict:
+        return Check.DEFAULT_CHECK_CONFIG
+
+    # @abstractmethod
+    # def parse_args(self, args: argparse) -> Dict:
+    #     raise NotImplementedError