+++ tests/checks/test_brackets.py
@@ -6,9 +6,10 @@
 # https://github.com/MarcinOrlowski/prop-tool/
 #
 """
-from proptool.checks.base.check import Check
+from typing import Union
+
 from proptool.checks.brackets import Brackets
-from proptool.config import Config
+from proptool.config.config import Config
 from proptool.decorators.overrides import overrides
 from proptool.prop.items import Blank, Comment, Translation
 from tests.checks.checks_test_case import ChecksTestCase
@@ -17,7 +18,7 @@ from tests.checks.checks_test_case import ChecksTestCase
 class ChecksBrackets(ChecksTestCase):
 
     @overrides(ChecksTestCase)
-    def get_checker(self, config: Config) -> Check:
+    def get_checker(self, config: Union[Config, None] = None) -> Brackets:
         return Brackets(config)
 
     # #################################################################################################
@@ -58,15 +59,16 @@ class ChecksBrackets(ChecksTestCase):
         """
         Checks if lists defining opening and closing markers are sane.
         """
-        checker: Brackets = self.get_checker(self.config)
+        checker: Brackets = self.get_checker(None)
+        config = checker.get_default_config()
 
-        self.assertEqual(len(checker.opening), len(checker.closing))
-        self.assertNotEqual([], checker.opening)
-        self.assertNotEqual([], checker.closing)
-        self.assertNotEqual(checker.opening, checker.closing)
+        self.assertEqual(len(config['opening']), len(config['closing']))
+        self.assertNotEqual([], config['opening'])
+        self.assertNotEqual([], config['closing'])
+        self.assertNotEqual(config['opening'], config['closing'])
 
         # ensure no marker is in both lists
-        for op_idx, op_marker in enumerate(checker.opening):
-            self.assertFalse(op_marker in checker.closing, f'Marker {op_marker} (position: {op_idx}) is present in closing too.')
-        for cl_idx, cl_marker in enumerate(checker.closing):
-            self.assertFalse(cl_marker in checker.opening, f'Marker {cl_marker} (position: {cl_idx}) is present in opening too.')
+        for op_idx, op_marker in enumerate(config['opening']):
+            self.assertFalse(op_marker in config['closing'], f'Marker {op_marker} (position: {op_idx}) is present in closing too.')
+        for cl_idx, cl_marker in enumerate(config['closing']):
+            self.assertFalse(cl_marker in config['opening'], f'Marker {cl_marker} (position: {cl_idx}) is present in opening too.')