+++ tests/checks/test_punctuation.py
@@ -7,10 +7,11 @@
 #
 """
 import random
+from typing import Union
 
 from proptool.checks.base.check import Check
 from proptool.checks.punctuation import Punctuation
-from proptool.config import Config
+from proptool.config.config import Config
 from proptool.decorators.overrides import overrides
 from proptool.prop.file import PropFile
 from proptool.prop.items import Translation
@@ -20,7 +21,7 @@ from tests.checks.checks_test_case import ChecksTestCase
 class TestPunctuation(ChecksTestCase):
 
     @overrides(ChecksTestCase)
-    def get_checker(self, config: Config) -> Check:
+    def get_checker(self, config: Union[Config, None] = None) -> Check:
         return Punctuation(config)
 
     # #################################################################################################