+++ tests/checks/test_trailing_white_chars.py
@@ -6,9 +6,11 @@
 # https://github.com/MarcinOrlowski/prop-tool/
 #
 """
+from typing import Union
+
 from proptool.checks.base.check import Check
 from proptool.checks.trailing_white_chars import TrailingWhiteChars
-from proptool.config import Config
+from proptool.config.config import Config
 from proptool.decorators.overrides import overrides
 from proptool.prop.items import Comment, Translation
 from tests.checks.checks_test_case import ChecksTestCase
@@ -17,7 +19,7 @@ from tests.checks.checks_test_case import ChecksTestCase
 class TestTrailingWhiteChars(ChecksTestCase):
 
     @overrides(ChecksTestCase)
-    def get_checker(self, config: Config) -> Check:
+    def get_checker(self, config: Union[Config, None] = None) -> Check:
         return TrailingWhiteChars(config)
 
     # #################################################################################################