+++ tests/checks/test_typesetting_quotation_marks.py
@@ -6,9 +6,10 @@
 # https://github.com/MarcinOrlowski/prop-tool/
 #
 """
-from proptool.checks.base.check import Check
+from typing import Union
+
 from proptool.checks.typesetting_quotation_marks import TypesettingQuotationMarks
-from proptool.config import Config
+from proptool.config.config import Config
 from proptool.decorators.overrides import overrides
 from proptool.prop.items import Blank, Comment, Translation
 from tests.checks.checks_test_case import ChecksTestCase
@@ -17,7 +18,7 @@ from tests.checks.checks_test_case import ChecksTestCase
 class ChecksBrackets(ChecksTestCase):
 
     @overrides(ChecksTestCase)
-    def get_checker(self, config: Config) -> Check:
+    def get_checker(self, config: Union[Config, None] = None) -> TypesettingQuotationMarks:
         return TypesettingQuotationMarks(config)
 
     # #################################################################################################